MGMT-8367 - Fix download_logs package. Fix vip_dhcp_allocation default value.
