+++ scripts/download_logs.sh
@@ -48,7 +48,7 @@ function download_cluster_logs() {
     fi
   fi
 
-  skipper run -e JUNIT_REPORT_DIR ./src/assisted_test_infra/download_logs/download_logs.py ${SERVICE_URL} ${LOGS_DEST} --cluster-id ${CLUSTER_ID} ${ADDITIONAL_PARAMS}
+  skipper run -e JUNIT_REPORT_DIR python3 -m src.assisted_test_infra.download_logs ${SERVICE_URL} ${LOGS_DEST} --cluster-id ${CLUSTER_ID} ${ADDITIONAL_PARAMS}
 }
 
 function collect_kube_api_resources() {
