+++ src/assisted_test_infra/download_logs/__init__.py
@@ -1,3 +1,3 @@
-from .download_logs import download_must_gather, gather_sosreport_data
+from .download_logs import download_must_gather, gather_sosreport_data, collect_debug_info_from_cluster
 
-__all__ = ["download_must_gather", "gather_sosreport_data"]
+__all__ = ["download_must_gather", "gather_sosreport_data", "collect_debug_info_from_cluster"]
