+++ src/assisted_test_infra/download_logs/download_logs.py
@@ -6,8 +6,6 @@ import os
 import shutil
 import tempfile
 import time
-from argparse import ArgumentParser
-from collections import Counter
 from contextlib import suppress
 from datetime import datetime
 
@@ -20,7 +18,7 @@ from paramiko.ssh_exception import SSHException
 from scp import SCPException
 
 from assisted_test_infra.test_infra.logger import log, add_log_file_handler, SuppressAndLog
-from assisted_test_infra.test_infra.assisted_service_api import InventoryClient, ClientFactory
+from assisted_test_infra.test_infra.assisted_service_api import InventoryClient
 from assisted_test_infra.test_infra.consts import ClusterStatus, HostsProgressStages, env_defaults
 from assisted_test_infra.test_infra.controllers.node_controllers.libvirt_controller import \
     LibvirtController
@@ -28,8 +26,7 @@ from assisted_test_infra.test_infra.controllers.node_controllers.node import Nod
 from assisted_test_infra.test_infra.helper_classes import cluster as helper_cluster
 from assisted_test_infra.test_infra.tools.concurrently import run_concurrently
 from assisted_test_infra.test_infra.utils import (are_host_progress_in_stage, config_etc_hosts,
-                              fetch_url, recreate_folder, run_command,
-                              verify_logs_uploaded, get_env)
+                              fetch_url, recreate_folder, run_command, verify_logs_uploaded)
 from tests.config import ClusterConfig, TerraformConfig
 
 private_ssh_key_path_default = os.path.join(os.getcwd(), str(env_defaults.DEFAULT_SSH_PRIVATE_KEY_PATH))
@@ -38,40 +35,12 @@ TIME_FORMAT = '%Y-%m-%d_%H:%M:%S'
 MAX_RETRIES = 3
 MUST_GATHER_MAX_RETRIES = 15
 RETRY_INTERVAL = 60 * 5
-CONNECTION_TIMEOUT = 30
 SOSREPORT_SCRIPT = os.path.join(
     os.path.dirname(os.path.abspath(__file__)),
     "resources",
     "man_sosreport.sh")
 
 urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
-
-
-def main():
-    args = handle_arguments()
-
-    if args.sosreport:
-        gather_sosreport_data(output_dir=args.dest)
-
-    client = ClientFactory.create_client(url=args.inventory_url, timeout=CONNECTION_TIMEOUT,
-                                         offline_token=get_env("OFFLINE_TOKEN"))
-    if args.cluster_id:
-        cluster = client.cluster_get(args.cluster_id)
-        download_cluster_logs(client, json.loads(json.dumps(cluster.to_dict(), sort_keys=True, default=str)), args.dest,
-                              args.must_gather, args.update_by_events, pull_secret=args.pull_secret)
-    else:
-        clusters = get_clusters(client, args.download_all)
-
-        if not clusters:
-            log.info('No clusters were found')
-            return
-
-        for cluster in clusters:
-            if args.download_all or should_download_logs(cluster):
-                download_cluster_logs(client, cluster, args.dest, args.must_gather, args.update_by_events,
-                                      pull_secret=args.pull_secret)
-
-        log.info("Cluster installation statuses: %s", dict(Counter(cluster["status"] for cluster in clusters).items()))
 
 
 def download_cluster_logs(client: InventoryClient, cluster: dict, dest: str, must_gather: bool,
@@ -309,22 +278,3 @@ def fetch_url_and_write_to_file(url_key, file_name, debug_info, output_folder):
             _file.write(content)
     else:
         log.warning(f"{url_key} is not available")
-
-
-def handle_arguments():
-    parser = ArgumentParser(description="Download logs")
-
-    parser.add_argument("inventory_url", help="URL of remote inventory", type=str)
-    parser.add_argument("dest", help="Destination to download logs", type=str)
-    parser.add_argument("--cluster-id", help="Cluster id to download its logs", type=str, default=None, nargs='?')
-    parser.add_argument("--download-all", help="Download logs from all clusters", action='store_true')
-    parser.add_argument("--must-gather", help="must-gather logs", action='store_true')
-    parser.add_argument("--sosreport", help="gather sosreport from each node", action='store_true')
-    parser.add_argument("--update-by-events", help="Update logs if cluster events were updated", action='store_true')
-    parser.add_argument("-ps", "--pull-secret", help="Pull secret", type=str, default="")
-
-    return parser.parse_args()
-
-
-if __name__ == '__main__':
-    main()
