+++ src/assisted_test_infra/test_infra/consts/env_defaults.py
@@ -5,6 +5,7 @@ from assisted_test_infra.test_infra import consts
 DEFAULT_NUMBER_OF_MASTERS: int = consts.NUMBER_OF_MASTERS
 DEFAULT_DAY2_WORKERS_COUNT: int = 1
 DEFAULT_WORKERS_COUNT: int = 0
+DEFAULT_VIP_DHCP_ALLOCATION: bool = True
 DEFAULT_STORAGE_POOL_PATH: Path = Path.cwd().joinpath("storage_pool")
 DEFAULT_SSH_PRIVATE_KEY_PATH: Path = Path.home() / ".ssh" / "id_rsa"
 DEFAULT_SSH_PUBLIC_KEY_PATH: Path = Path.home() / ".ssh" / "id_rsa.pub"
