+++ src/assisted_test_infra/test_infra/controllers/node_controllers/libvirt_controller.py
@@ -15,8 +15,8 @@ import libvirt
 import waiting
 
 from assisted_test_infra.test_infra import BaseClusterConfig, BaseInfraEnvConfig, consts, utils
-from assisted_test_infra.test_infra.controllers.node_controllers.node import Node
 from assisted_test_infra.test_infra.controllers.node_controllers.disk import Disk, DiskSourceType
+from assisted_test_infra.test_infra.controllers.node_controllers.node import Node
 from assisted_test_infra.test_infra.controllers.node_controllers.node_controller import NodeController
 from assisted_test_infra.test_infra.helper_classes.config.controller_config import BaseNodeConfig
 from assisted_test_infra.test_infra.logger import log
