+++ src/assisted_test_infra/test_infra/controllers/node_controllers/terraform_controller.py
@@ -19,8 +19,8 @@ from assisted_test_infra.test_infra import (
     virsh_cleanup,
 )
 from assisted_test_infra.test_infra.consts import resources
-from assisted_test_infra.test_infra.controllers.node_controllers.node import Node
 from assisted_test_infra.test_infra.controllers.node_controllers.libvirt_controller import LibvirtController
+from assisted_test_infra.test_infra.controllers.node_controllers.node import Node
 from assisted_test_infra.test_infra.tools import static_network, terraform_utils
 from assisted_test_infra.test_infra.utils import TerraformControllerUtil
 from assisted_test_infra.test_infra.utils.base_name import BaseName, get_name_suffix
