+++ src/assisted_test_infra/test_infra/utils/__init__.py
@@ -1,6 +1,6 @@
 from .logs_utils import verify_logs_uploaded
 from .terraform_util import TerraformControllerUtil
 from .utils import *  # TODO - temporary import all old utils
-from .utils import get_env
+from .utils import get_env, get_openshift_release_image
 
-__all__ = ["verify_logs_uploaded", "get_env", "TerraformControllerUtil"]
+__all__ = ["verify_logs_uploaded", "get_env", "TerraformControllerUtil", "get_openshift_release_image"]
